+++ Pod/Classes/CZPickerView.m
@@ -29,6 +29,7 @@ typedef void (^CZDismissCompletionCallback)(void);
 @property UIView *footerview;
 @property UITableView *tableView;
 @property NSMutableArray *selectedIndexPaths;
+@property CGRect previousBounds;
 @end
 
 @implementation CZPickerView
@@ -64,8 +65,8 @@ typedef void (^CZDismissCompletionCallback)(void);
         self.confirmButtonHighlightedColor = [UIColor colorWithRed:236.0/255 green:240/255.0 blue:241.0/255 alpha:1];
         self.confirmButtonBackgroundColor = [UIColor colorWithRed:56.0/255 green:185.0/255 blue:158.0/255 alpha:1];
         
-        CGRect rect= [UIScreen mainScreen].bounds;
-        self.frame = rect;
+        _previousBounds = [UIScreen mainScreen].bounds;
+        self.frame = _previousBounds;
     }
     return self;
 }
@@ -158,7 +159,8 @@ typedef void (^CZDismissCompletionCallback)(void);
 }
 
 - (UIView *)buildContainerView{
-    CGAffineTransform transform = CGAffineTransformMake(0.8, 0, 0, 0.8, 0, 0);
+    CGFloat widthRatio = _pickerWidth ? _pickerWidth / [UIScreen mainScreen].bounds.size.width : 0.8;
+    CGAffineTransform transform = CGAffineTransformMake(widthRatio, 0, 0, 0.8, 0, 0);
     CGRect newRect = CGRectApplyAffineTransform(self.frame, transform);
     UIView *cv = [[UIView alloc] initWithFrame:newRect];
     cv.layer.cornerRadius = 6.0f;
@@ -168,7 +170,8 @@ typedef void (^CZDismissCompletionCallback)(void);
 }
 
 - (UITableView *)buildTableView{
-    CGAffineTransform transform = CGAffineTransformMake(0.8, 0, 0, 0.8, 0, 0);
+    CGFloat widthRatio = _pickerWidth ? _pickerWidth / [UIScreen mainScreen].bounds.size.width : 0.8;
+    CGAffineTransform transform = CGAffineTransformMake(widthRatio, 0, 0, 0.8, 0, 0);
     CGRect newRect = CGRectApplyAffineTransform(self.frame, transform);
     NSInteger n = [self.dataSource numberOfRowsInPickerView:self];
     CGRect tableRect;
@@ -248,9 +251,12 @@ typedef void (^CZDismissCompletionCallback)(void);
 - (UIView *)buildHeaderView{
     UIView *view = [[UIView alloc] initWithFrame:CGRectMake(0, 0, self.tableView.frame.size.width, CZP_HEADER_HEIGHT)];
     view.backgroundColor = self.headerBackgroundColor;
+    
+    UIFont *headerFont = self.headerTitleFont == nil ? [UIFont systemFontOfSize:18.0] : self.headerTitleFont;
+    
     NSDictionary *dict = @{
                            NSForegroundColorAttributeName: self.headerTitleColor,
-                           NSFontAttributeName: [UIFont systemFontOfSize:18.0]
+                           NSFontAttributeName:headerFont
                            };
     NSAttributedString *at = [[NSAttributedString alloc] initWithString:self.headerTitle attributes:dict];
     UILabel *label = [[UILabel alloc] initWithFrame:view.frame];
@@ -339,6 +345,10 @@ typedef void (^CZDismissCompletionCallback)(void);
     } else if([self.dataSource respondsToSelector:@selector(czpickerView:titleForRow:)]){
         cell.textLabel.text = [self.dataSource czpickerView:self titleForRow:indexPath.row];
     }
+    
+    if(self.checkmarkColor){
+        cell.tintColor = self.checkmarkColor;
+    }
     return cell;
 }
 
@@ -397,52 +407,25 @@ typedef void (^CZDismissCompletionCallback)(void);
     NSMutableSet *set = [NSMutableSet set];
     for(NSString *o in supportedOrientations){
         NSRange range = [o rangeOfString:@"Portrait"];
-        if ( range.location != NSNotFound ) {
+        if (range.location != NSNotFound) {
             [set addObject:@"Portrait"];
         }
         
         range = [o rangeOfString:@"Landscape"];
-        if ( range.location != NSNotFound ) {
+        if (range.location != NSNotFound) {
             [set addObject:@"Landscape"];
         }
     }
     return set.count == 2;
 }
 
-- (BOOL)orientationSupported:(UIDeviceOrientation)orientation{
-    NSString *orientationStr;
-    switch (orientation) {
-        case UIDeviceOrientationPortrait:
-            orientationStr = @"UIInterfaceOrientationPortrait";
-            break;
-        case UIDeviceOrientationPortraitUpsideDown:
-            orientationStr = @"UIInterfaceOrientationPortraitUpsideDown";
-            break;
-        case UIDeviceOrientationLandscapeLeft:
-            orientationStr = @"UIInterfaceOrientationLandscapeLeft";
-            break;
-        case UIDeviceOrientationLandscapeRight:
-            orientationStr = @"UIInterfaceOrientationLandscapeRight";
-            break;
-        default:
-            orientationStr = @"Invalid Interface Orientation";
-            break;
-    }
-    NSArray *supportedOrientations = [[[NSBundle mainBundle] infoDictionary]
-                                      objectForKey:@"UISupportedInterfaceOrientations"];
-    for(NSString *o in supportedOrientations){
-        if([o hasPrefix:orientationStr]){
-            return YES;
-        }
-    }
-    return NO;
-}
-
 - (void)deviceOrientationDidChange:(NSNotification *)notification{
-    if(![self orientationSupported:[[UIDevice currentDevice] orientation]]){
+    CGRect rect = [UIScreen mainScreen].bounds;
+    if (CGRectEqualToRect(rect, _previousBounds)) {
         return;
     }
-    self.frame = [UIScreen mainScreen].bounds;
+    _previousBounds = rect;
+    self.frame = rect;
     for(UIView *v in self.subviews){
         if([v isEqual:self.backgroundDimmingView]) continue;